adds contactSchema and delete "contacts" folder
