--- model/contacts.js
@@ -1,23 +1,0 @@
-const { Schema, model } = require('mongoose')
-const Joi = require('joi')
-
-const contactSchema = Schema({
-  name: {
-    type: String,
-    required: [true, 'Set name for contact'],
-  },
-  email: {
-    type: String,
-  },
-  phone: {
-    type: String,
-  },
-  favorite: {
-    type: Boolean,
-    default: false,
-  },
-})
-
-const Contact = model('contact', contactSchema)
-
-module.exports = Contact
