--- model/contacts/index.js
@@ -1,13 +1,0 @@
-const listContacts = require('./listContacts')
-const getContactById = require('./getContactById')
-const addContact = require('./addContact')
-const removeContactById = require('./removeContactById')
-const updateContactById = require('./updateContactById')
-
-module.exports = {
-  listContacts,
-  getContactById,
-  addContact,
-  removeContactById,
-  updateContactById,
-}
