--- model/contacts/listContacts.js
@@ -1,8 +1,0 @@
-const getAllContacts = require('./getAllContacts')
-
-const listContacts = async () => {
-  const contacts = await getAllContacts()
-  return contacts
-}
-
-module.exports = listContacts
