--- model/contacts/updateContacts.js
@@ -1,7 +1,0 @@
-
-const fs = require('fs/promises')
-const { contactsPath } = require('../../utils')
-
-const updateContacts = async (newContact) => fs.writeFile(contactsPath, JSON.stringify(newContact))
-
-module.exports = updateContacts
